+++ plugins/gcp-projects/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-gcp-projects
 
+## 0.3.43-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.13.8-next.1
+  - @backstage/core-plugin-api@1.8.0-next.0
+  - @backstage/theme@0.4.4-next.0
+
 ## 0.3.43-next.0
 
 ### Patch Changes